+++ sdl2/pdcdisp.c
@@ -5,89 +5,99 @@
 #include <stdlib.h>
 #include <string.h>
 
-# ifdef PDC_WIDE
-#  define USE_UNICODE_ACS_CHARS 1
-# else
-#  define USE_UNICODE_ACS_CHARS 0
-# endif
+#define USE_UNICODE_ACS_CHARS 1
 
 # include "../common/acs_defs.h"
 # include "../common/pdccolor.h"
 
-#define MAXRECT 200     /* maximum number of rects to queue up before
-                           an update is forced; the number was chosen
-                           arbitrarily */
-
-static SDL_Rect uprect[MAXRECT];       /* table of rects to update */
 static chtype oldch = (chtype)(-1);    /* current attribute */
-static int rectcount = 0;              /* index into uprect */
 static int foregr = -2, backgr = -2; /* current foreground, background */
+static int cache_attr_index = 0;
 static bool blinked_off = FALSE;
-/* do the real updates on a delay */
 
-void PDC_update_rects(void)
+static SDL_Color get_pdc_color( const int color_idx)
 {
-    int i;
-
-    if (rectcount)
-    {
-        /* if the maximum number of rects has been reached, we're
-           probably better off doing a full screen update */
-
-        if (rectcount == MAXRECT)
-            SDL_UpdateWindowSurface(pdc_window);
-        else
-        {
-            int w = pdc_screen->w;
-            int h = pdc_screen->h;
-
-            for (i = 0; i < rectcount; i++)
-            {
-                if (uprect[i].x > w ||
-                    uprect[i].y > h ||
-                    !uprect[i].w || !uprect[i].h)
-                {
-                    if (i + 1 < rectcount)
-                    {
-                        memmove(uprect + i, uprect + i + 1,
-                                (rectcount - i + 1) * sizeof(*uprect));
-                        --i;
-                    }
-                    rectcount--;
-                    continue;
-                }
-
-                if (uprect[i].x + uprect[i].w > w)
-                    uprect[i].w = min(w, w - uprect[i].x);
-
-                if (uprect[i].y + uprect[i].h > h)
-                    uprect[i].h = min(h, h - uprect[i].y);
-            }
-
-            if (rectcount > 0)
-                SDL_UpdateWindowSurfaceRects(pdc_window, uprect, rectcount);
-        }
-
-        rectcount = 0;
-    }
-}
-
-static SDL_Color *get_pdc_color( const int color_idx)
-{
-    static SDL_Color c;
+    SDL_Color c;
     const PACKED_RGB rgb = PDC_get_palette_entry( color_idx > 0 ? color_idx : 0);
 
     c.r = (Uint8)Get_RValue( rgb);
     c.g = (Uint8)Get_GValue( rgb);
     c.b = (Uint8)Get_BValue( rgb);
-    return( &c);
+    return c;
 }
 
-static Uint32 get_pdc_mapped( const int color_idx)
+static SDL_Texture* get_glyph_texture(Uint32 ch32)
 {
-   SDL_Color *c = get_pdc_color( color_idx);
+    SDL_Texture* glyph = NULL;
+    SDL_Color color = get_pdc_color(foregr);
+    SDL_Color white = {255,255,255,255};
 
-   return( SDL_MapRGB( pdc_screen->format, c->r, c->g, c->b));
+    int *cache_size = &pdc_glyph_cache_size[cache_attr_index];
+    SDL_Texture*** cache = &pdc_glyph_cache[cache_attr_index];
+
+    if(ch32 < *cache_size && (*cache)[ch32] != NULL)
+    {
+        glyph = (*cache)[ch32];
+    }
+    else
+    {
+        SDL_Surface* surf = NULL;
+
+#ifdef PDC_SDL_SUPPLEMENTARY_PLANES_SUPPORT
+        switch (pdc_sdl_render_mode)
+        {
+        case PDC_SDL_RENDER_SOLID:
+            surf = TTF_RenderGlyph32_Solid(pdc_ttffont, ch32, white);
+            break;
+        default:
+            surf = TTF_RenderGlyph32_Blended(pdc_ttffont, ch32, white);
+        }
+#else
+        /* no support for supplementary planes */
+        if (ch > 0xffff)
+            ch = '?';
+
+        switch (pdc_sdl_render_mode)
+        {
+        case PDC_SDL_RENDER_SOLID:
+            surf = TTF_RenderGlyph_Solid(pdc_ttffont, (Uint16)ch32, white);
+            break;
+        default:
+            surf = TTF_RenderGlyph_Blended(pdc_ttffont, (Uint16)ch32, white);
+        }
+#endif
+        glyph = SDL_CreateTextureFromSurface(pdc_renderer, surf);
+        SDL_FreeSurface(surf);
+
+        if(ch32 >= *cache_size)
+        {
+            int new_cache_size = *cache_size;
+            if(new_cache_size == 0) new_cache_size = 256;
+            while(new_cache_size < ch32) (new_cache_size) *= 2;
+
+            *cache = realloc(*cache, sizeof(SDL_Texture*)*new_cache_size);
+            memset(
+                (*cache) + *cache_size, 0,
+                sizeof(SDL_Texture*)*(new_cache_size - *cache_size)
+            );
+            *cache_size = new_cache_size;
+        }
+        (*cache)[ch32] = glyph;
+    }
+
+    SDL_SetTextureColorMod(glyph, color.r, color.g, color.b);
+    return glyph;
+}
+
+static void fill_rect(SDL_Rect* rect, SDL_Color col)
+{
+    SDL_SetRenderDrawColor(pdc_renderer, col.r, col.g, col.b, col.a);
+    SDL_RenderFillRect(pdc_renderer, rect);
+}
+
+static void blit_texture(SDL_Texture* glyph, SDL_Rect* src, SDL_Rect* dest)
+{
+    SDL_RenderCopy(pdc_renderer, glyph, src, dest);
 }
 
 /* set the font colors to match the chtype's attribute */
@@ -96,13 +106,13 @@ static void _set_attr(chtype ch)
 {
     attr_t sysattrs = SP->termattrs;
 
-#ifdef PDC_WIDE
-    TTF_SetFontStyle(pdc_ttffont,
-        ( ((ch & A_BOLD) && (sysattrs & A_BOLD)) ?
-            TTF_STYLE_BOLD : 0) |
-        ( ((ch & A_ITALIC) && (sysattrs & A_ITALIC)) ?
-            TTF_STYLE_ITALIC : 0) );
-#endif
+    int bold = (ch & A_BOLD) && (sysattrs & A_BOLD);
+    int italic = (ch & A_ITALIC) && (sysattrs & A_ITALIC);
+    cache_attr_index = (bold ? 1 : 0) | (italic ? 2 : 0);
+    TTF_SetFontStyle(
+        pdc_ttffont,
+        (bold ? TTF_STYLE_BOLD : 0) | (italic ? TTF_STYLE_ITALIC : 0)
+    );
 
     ch &= (A_COLOR|A_BOLD|A_BLINK|A_REVERSE);
 
@@ -129,27 +139,11 @@ static void _set_attr(chtype ch)
 
         if (newfg != foregr)
         {
-#ifndef PDC_WIDE
-            SDL_SetPaletteColors(pdc_font->format->palette,
-                                 get_pdc_color( newfg), pdc_flastc, 1);
-#endif
             foregr = newfg;
         }
 
         if (newbg != backgr)
         {
-#ifndef PDC_WIDE
-            if (newbg == -1)
-                SDL_SetColorKey(pdc_font, SDL_TRUE, 0);
-            else
-            {
-                if (backgr == -1)
-                    SDL_SetColorKey(pdc_font, SDL_FALSE, 0);
-
-                SDL_SetPaletteColors(pdc_font->format->palette,
-                                     get_pdc_color( newbg), 0, 1);
-            }
-#endif
             backgr = newbg;
         }
 
@@ -157,7 +151,6 @@ static void _set_attr(chtype ch)
     }
 }
 
-#ifdef PDC_WIDE
 /* Draw some of the ACS_* "graphics" */
 
 #define BIT_UP       1
@@ -192,10 +185,10 @@ static bool _grprint(chtype ch, const SDL_Rect dest)
         const int wmid = (pdc_fwidth - pdc_fthick) >> 1;
         const int mask = remap_tbl[i + 1];
         SDL_Rect temp = dest;
-        const Uint32 col = get_pdc_mapped( foregr);
+        const SDL_Color col = get_pdc_color( foregr);
 
         if( ch == ACS_BLOCK)
-            SDL_FillRect(pdc_screen, &temp, col);
+            fill_rect(&temp, col);
         if( mask & HORIZ)
         {
             temp.h = pdc_fthick;
@@ -215,7 +208,7 @@ static bool _grprint(chtype ch, const SDL_Rect dest)
                 case HORIZ:
                     break;
             }
-            SDL_FillRect(pdc_screen, &temp, col);
+            fill_rect(&temp, col);
         }
         temp = dest;
         if( mask & VERTIC)
@@ -234,7 +227,7 @@ static bool _grprint(chtype ch, const SDL_Rect dest)
                 case VERTIC:
                     break;
             }
-            SDL_FillRect(pdc_screen, &temp, col);
+            fill_rect(&temp, col);
         }
         rval = TRUE;
     }
@@ -242,48 +235,6 @@ static bool _grprint(chtype ch, const SDL_Rect dest)
         rval = FALSE;   /* didn't draw it -- fall back to acs_map */
     return( rval);
 }
-
-static SDL_Surface *_render_ttf_glyph(chtype ch)
-{
-    ch &= A_CHARTEXT;
-
-#ifdef PDC_SDL_SUPPLEMENTARY_PLANES_SUPPORT
-    switch (pdc_sdl_render_mode)
-    {
-    case PDC_SDL_RENDER_SOLID:
-        return TTF_RenderGlyph32_Solid(pdc_ttffont, (Uint32)ch,
-                                       *(get_pdc_color( foregr)));
-    case PDC_SDL_RENDER_SHADED:
-        return TTF_RenderGlyph32_Shaded(pdc_ttffont, (Uint32)ch,
-                                        *(get_pdc_color( foregr)),
-                                        *(get_pdc_color( backgr)));
-    default:
-        return TTF_RenderGlyph32_Blended(pdc_ttffont, (Uint32)ch,
-                                         *(get_pdc_color( foregr)));
-    }
-#else
-    /* no support for supplementary planes */
-    if (ch > 0xffff)
-        ch = '?';
-
-    switch (pdc_sdl_render_mode)
-    {
-    case PDC_SDL_RENDER_SOLID:
-        return TTF_RenderGlyph_Solid(pdc_ttffont, (Uint16)ch,
-                                     *(get_pdc_color( foregr)));
-        break;
-    case PDC_SDL_RENDER_SHADED:
-        return TTF_RenderGlyph_Shaded(pdc_ttffont, (Uint16)ch,
-                                      *(get_pdc_color( foregr)),
-                                      *(get_pdc_color( backgr)));
-    default:
-        return TTF_RenderGlyph_Blended(pdc_ttffont, (Uint16)ch,
-                                       *(get_pdc_color( foregr)));
-    }
-#endif
-}
-
-#endif
 
 /* draw a cursor at (y, x) */
 
@@ -305,7 +256,7 @@ void PDC_gotoyx(int row, int col)
 
     if (!SP->visibility)
     {
-        PDC_update_rects();
+        //PDC_doupdate();
         return;
     }
 
@@ -325,8 +276,7 @@ void PDC_gotoyx(int row, int col)
     dest.h = src.h;
     dest.w = src.w;
 
-#ifdef PDC_WIDE
-    SDL_FillRect(pdc_screen, &dest, get_pdc_mapped( backgr));
+    fill_rect(&dest, get_pdc_color( backgr));
 
     if (!(SP->visibility == 2 && _is_altcharset( ch) &&
         _grprint(ch & (0x7f | A_ALTCHARSET), dest)))
@@ -334,39 +284,25 @@ void PDC_gotoyx(int row, int col)
         if( _is_altcharset( ch))
             ch = acs_map[ch & 0x7f];
 
-        pdc_font = _render_ttf_glyph(ch);
+        Uint32 ch32 = (Uint32)(ch & A_CHARTEXT);
 
-        if (pdc_font)
+        SDL_Texture* glyph = get_glyph_texture(ch32);
+
+        if (glyph)
         {
-            int center = pdc_fwidth > pdc_font->w ?
-                        (pdc_fwidth - pdc_font->w) >> 1 : 0;
+            int w = 0;
+            SDL_QueryTexture(glyph, NULL, NULL, &w, NULL);
+            int center = pdc_fwidth > w ? (pdc_fwidth - w) >> 1 : 0;
             src.x = 0;
             src.y = pdc_fheight - src.h;
             dest.x += center;
-            SDL_BlitSurface(pdc_font, &src, pdc_screen, &dest);
+            blit_texture(glyph, &src, &dest);
             dest.x -= center;
-            SDL_FreeSurface(pdc_font);
-            pdc_font = NULL;
+            glyph = NULL;
         }
     }
-#else
-    if( _is_altcharset( ch))
-        ch = acs_map[ch & 0x7f];
 
-    src.x = (ch & 0xff) % 32 * pdc_fwidth;
-    src.y = (ch & 0xff) / 32 * pdc_fheight + (pdc_fheight - src.h);
-
-    SDL_BlitSurface(pdc_font, &src, pdc_screen, &dest);
-#endif
-
-    if (oldrow != row || oldcol != col)
-    {
-        if (rectcount == MAXRECT)
-            PDC_update_rects();
-
-        uprect[rectcount++] = dest;
-    }
-    PDC_update_rects();
+    //PDC_doupdate();
 }
 
 static bool _merge_rects( SDL_Rect *a, const SDL_Rect *b)
@@ -408,20 +344,13 @@ void _new_packet(attr_t attr, int lineno, int x, int len, const chtype *srcp)
 {
     SDL_Rect src, dest;
     int j;
-#ifdef PDC_WIDE
-    chtype prev_ch = 0;
-#endif
+    Uint32 prev_ch = 0;
     attr_t sysattrs = SP->termattrs;
     short hcol = SP->line_color;
     bool blink = blinked_off && (attr & A_BLINK) && (sysattrs & A_BLINK);
 
-    if (rectcount == MAXRECT)
-        PDC_update_rects();
-
-#ifdef PDC_WIDE
     src.x = 0;
     src.y = 0;
-#endif
     src.h = pdc_fheight;
     src.w = pdc_fwidth;
 
@@ -429,28 +358,15 @@ void _new_packet(attr_t attr, int lineno, int x, int len, const chtype *srcp)
     dest.x = pdc_fwidth * x + pdc_xoffset;
     dest.h = pdc_fheight;
     dest.w = pdc_fwidth * len;
-    uprect[rectcount++] = dest;
-
-    /* we try merging the new rectangle with the previous one.  It's
-       possible the result can be merged with the one before that,
-       and so on.  */
-
-    while( rectcount >= 2 && _merge_rects( uprect + rectcount - 2,
-                                           uprect + rectcount - 1))
-       rectcount--;
 
     _set_attr(attr);
 
-    if (backgr == -1)
-        SDL_BlitSurface(pdc_tileback, &dest, pdc_screen, &dest);
-#ifdef PDC_WIDE
-    else
-        SDL_FillRect(pdc_screen, &dest, get_pdc_mapped( backgr));
-#endif
+    fill_rect(&dest, get_pdc_color( backgr));
 
     if (hcol == -1)
         hcol = (short)foregr;
 
+    SDL_Texture* glyph = NULL;
     for (j = 0; j < len; j++)
     {
         chtype ch = srcp[j];
@@ -462,17 +378,14 @@ void _new_packet(attr_t attr, int lineno, int x, int len, const chtype *srcp)
 
         if( _is_altcharset( ch))
         {
-#ifdef PDC_WIDE
             if (_grprint(ch & (0x7f | A_ALTCHARSET), dest))
             {
                 dest.x += pdc_fwidth;
                 continue;
             }
-#endif
             ch = acs_map[ch & 0x7f];
         }
 
-#ifdef PDC_WIDE
         ch &= A_CHARTEXT;
 
         if (ch != ' ')
@@ -480,39 +393,32 @@ void _new_packet(attr_t attr, int lineno, int x, int len, const chtype *srcp)
             if (prev_ch != ch)
             {
                 prev_ch = ch;
-                if (pdc_font)
-                    SDL_FreeSurface(pdc_font);
 
-                pdc_font = _render_ttf_glyph(ch);
+                glyph = get_glyph_texture(prev_ch);
             }
 
-            if (pdc_font)
+            if (glyph)
             {
-                int center = pdc_fwidth > pdc_font->w ?
-                    (pdc_fwidth - pdc_font->w) >> 1 : 0;
+                int w = 0;
+                SDL_QueryTexture(glyph, NULL, NULL, &w, NULL);
+                int center = pdc_fwidth > w ? (pdc_fwidth - w) >> 1 : 0;
                 dest.x += center;
-                SDL_BlitSurface(pdc_font, &src, pdc_screen, &dest);
+                blit_texture(glyph, &src, &dest);
                 dest.x -= center;
             }
         }
-#else
-        src.x = (ch & 0xff) % 32 * pdc_fwidth;
-        src.y = (ch & 0xff) / 32 * pdc_fheight;
-
-        SDL_BlitSurface(pdc_font, &src, pdc_screen, &dest);
-#endif
 
         if (!blink && (attr & (A_LEFT | A_RIGHT)))
         {
             dest.w = pdc_fthick;
 
             if (attr & A_LEFT)
-                SDL_FillRect(pdc_screen, &dest, get_pdc_mapped( hcol));
+                fill_rect(&dest, get_pdc_color( hcol));
 
             if (attr & A_RIGHT)
             {
                 dest.x += pdc_fwidth - pdc_fthick;
-                SDL_FillRect(pdc_screen, &dest, get_pdc_mapped( hcol));
+                fill_rect(&dest, get_pdc_color( hcol));
                 dest.x -= pdc_fwidth - pdc_fthick;
             }
         }
@@ -520,31 +426,23 @@ void _new_packet(attr_t attr, int lineno, int x, int len, const chtype *srcp)
         dest.x += pdc_fwidth;
     }
 
-#ifdef PDC_WIDE
-    if (pdc_font)
-    {
-        SDL_FreeSurface(pdc_font);
-        pdc_font = NULL;
-    }
-#endif
-
     if (!blink && (attr & (A_UNDERLINE | A_OVERLINE | A_STRIKEOUT)))
     {
         dest.x = pdc_fwidth * x + pdc_xoffset;
         dest.h = pdc_fthick;
         dest.w = pdc_fwidth * len;
         if( attr & A_OVERLINE)
-           SDL_FillRect(pdc_screen, &dest, get_pdc_mapped( hcol));
+           fill_rect(&dest, get_pdc_color( hcol));
         if( attr & A_UNDERLINE)
         {
            dest.y += pdc_fheight - pdc_fthick;
-           SDL_FillRect(pdc_screen, &dest, get_pdc_mapped( hcol));
+           fill_rect(&dest, get_pdc_color( hcol));
            dest.y -= pdc_fheight - pdc_fthick;
         }
         if( attr & A_STRIKEOUT)
         {
            dest.y += (pdc_fheight - pdc_fthick) / 2;
-           SDL_FillRect(pdc_screen, &dest, get_pdc_mapped( hcol));
+           fill_rect(&dest, get_pdc_color( hcol));
         }
     }
 }
@@ -630,7 +528,11 @@ void PDC_blink_text(void)
 
 void PDC_doupdate(void)
 {
-    PDC_update_rects();
+    SDL_SetRenderTarget(pdc_renderer, NULL);
+    SDL_RenderClear(pdc_renderer);
+    SDL_RenderCopy(pdc_renderer, pdc_render_target, NULL, NULL);
+    SDL_RenderPresent(pdc_renderer);
+    SDL_SetRenderTarget(pdc_renderer, pdc_render_target);
 }
 
 void PDC_pump_and_peep(void)
@@ -644,8 +546,7 @@ void PDC_pump_and_peep(void)
              SDL_WINDOWEVENT_EXPOSED == event.window.event ||
              SDL_WINDOWEVENT_SHOWN == event.window.event))
         {
-            SDL_UpdateWindowSurface(pdc_window);
-            rectcount = 0;
+            PDC_doupdate();
         }
         else
             SDL_PushEvent(&event);
