--- sdl2ren/versions.mif
@@ -1,2 +1,0 @@
-SDLBASE = C:\SDL2-2.0.10
-TTFBASE = C:\SDL2_ttf-2.0.15
